backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
motor==3.6.0
pydantic==2.10.4
pydantic-settings==2.7.0
google-genai==1.0.0
httpx==0.28.1
python-dotenv==1.0.1
